# openapi.yaml
openapi: 3.1.0
info:
  title: Cadence API
  version: "1.0.0"
  summary: REST API for the Cadence training platform.
  description: >
    A REST API over the same data model that powers the Cadence app — activities
    and their 1 Hz streams, designed workouts, scheduling, athlete thresholds and
    zones, coaching, and gear. JSON over HTTPS, OAuth 2.0, predictable resource URLs.


    ## Authentication

    OAuth 2.0 authorization code flow. Send the access token as a bearer header on
    every request:


    `Authorization: Bearer cad_at_9f2a7c…`


    ## Pagination

    List endpoints are cursor-paginated. Pass `cursor` from the previous response's
    `next_cursor`, plus an optional `limit` (max 200).


    ## Query language (CQL)

    The `q` parameter on `GET /v1/activities` accepts the same query language as the
    Activities search bar — field comparisons (`>` `<` `>=` `<=` `=` `!=`), boolean
    `AND` / `OR`, and `ORDER BY`. Example:
    `sport = run AND tag ~ race AND distance > 42.2km`.
  contact:
    name: Cadence Developer Platform
    url: https://developers.cadence.cc
  license:
    name: Proprietary
servers:
  - url: https://api.cadence.cc
    description: Production
security:
  - oauth2: []
  - bearerAuth: []
...
paths:
...
  /v1/athletes/{id}:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes & zones]
      summary: Retrieve an athlete
      operationId: getAthlete
      responses:
        "200":
          description: Profile and thresholds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Athlete"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Athletes & zones]
      summary: Update thresholds
      operationId: updateAthlete
      description: >
        Updates profile and threshold values. Changing a threshold recomputes every
        zone set that references it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                age: { type: integer }
                weight_kg: { type: number, description: Used for the Avg Watts/kg stat on Activity Analysis. }
                ftp:
                  type: integer
                  description: >
                    Functional threshold power (W). A manually-entered value is
                    recorded as a new threshold_history entry unconditionally (no
                    outlier check - an explicit human declaration, not a computed
                    candidate) and ages out after threshold_window_days like any
                    other entry. A no-op if it matches the latest ledger entry
                    already, since this form resubmits every field on every save.
                critical_run_power: { type: integer, description: Critical running power (W). Same threshold_history semantics as ftp. }
                threshold_pace: { type: string, description: "mm:ss per km. Same threshold_history semantics as ftp." }
                threshold_window_days: { type: integer, description: See the User schema. }
                threshold_sanity_pct: { type: integer, description: See the User schema. }
                threshold_warning_days: { type: integer, enum: [14, 21, 28, 0], description: See the User schema. }
                lthr: { type: integer, description: Lactate threshold HR (bpm). }
                max_hr: { type: integer }
                resting_hr:
                  type: integer
                  description: >
                    Optional; only feeds the Karvonen HRR% stat, not a zone
                    reference threshold, so changing it never appears in
                    zones_recomputed.
      responses:
        "200":
          description: Updated athlete.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Athlete"
                  - type: object
                    properties:
                      zones_recomputed:
                        type: array
                        items: { type: string }
  /v1/athletes/{id}/thresholds:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes & zones]
      summary: Threshold summary
      operationId: getThresholds
      description: >
        Current + previous value and staleness for ftp, critical_run_power, and
        threshold_pace - what the dashboard's threshold widget reads. No recompute
        here; stale just means "will update on the next activity, or via a manual
        refresh."
      responses:
        "200":
          description: Threshold summary for all three fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ftp: { $ref: "#/components/schemas/ThresholdSummaryEntry" }
                  critical_run_power: { $ref: "#/components/schemas/ThresholdSummaryEntry" }
                  threshold_pace: { $ref: "#/components/schemas/ThresholdSummaryEntry" }
  /v1/athletes/{id}/threshold-history:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes & zones]
      summary: Threshold history ledger
      operationId: getThresholdHistory
      description: The full history for one field, most recent first.
      parameters:
        - name: field
          in: query
          required: true
          schema: { type: string, enum: [ftp, critical_run_power, threshold_pace] }
      responses:
        "200":
          description: The ledger for the requested field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  field: { type: string, enum: [ftp, critical_run_power, threshold_pace] }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ThresholdHistoryEntry" }
  /v1/athletes/{id}/threshold-suggestions:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes & zones]
      summary: Threshold suggestions
      operationId: listThresholdSuggestions
      description: >
        Actionable suggestions where the auto-derivation in GET
        /v1/athletes/{id}/thresholds falls short: a qualifying effort the sanity
        band rejected ("rejected" - a real breakthrough, or a recalibrated/corrupt
        sensor; the athlete decides), or a current value about to age out of
        threshold_window_days with nothing dismissing it ("upcoming_drop", or
        "race_will_refresh" when a booked race should refresh it naturally first).
        LTHR is deliberately excluded - it's a stable, manually-set value, and HR
        bests are too noisy (heat, drift, sensor spikes) for either suggestion
        type. Cached per athlete for 24 hours, invalidated on anything that could
        change the list (activity ingest/import, a threshold write, a setting
        change, a race change, or an accept/dismiss).
      responses:
        "200":
          description: Pending suggestions, rejected first, then upcoming_drop/race_will_refresh by days_left ascending.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ThresholdSuggestion" }
  /v1/athletes/{id}/threshold-suggestions/{suggestion_id}/accept:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
      - name: suggestion_id
        in: path
        required: true
        schema: { type: string }
        description: "A suggestion's own id, e.g. ftp:rejected:act_123 - stable for its lifetime."
    post:
      tags: [Athletes & zones]
      summary: Accept a rejected candidate
      operationId: acceptThresholdSuggestion
      description: >
        Accepts a "rejected" suggestion as real: records it as a genuine
        threshold_history entry and updates the athlete's cached profile value.
        The candidate activity bypasses the sanity band for this field from now
        on (even across a full history rebuild), so it won't be rejected again.
        404 if the suggestion no longer exists.
      responses:
        "200":
          description: The new current threshold_history entry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ThresholdHistoryEntry" }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Athletes & zones]
      summary: Undo an accept
      operationId: undoAcceptThresholdSuggestion
      description: >
        Removes the accept override and re-runs the normal windowed recompute -
        the candidate goes back to being sanity-filtered like any other activity.
      responses:
        "204":
          description: Reverted.
  /v1/athletes/{id}/threshold-suggestions/{suggestion_id}/dismiss:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
      - name: suggestion_id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [Athletes & zones]
      summary: Dismiss a suggestion
      operationId: dismissThresholdSuggestion
      description: >
        Silences this one specific occurrence - a candidate activity id for
        "rejected", or a source-activity+expiry-date pair for "upcoming_drop" - so
        a genuinely different future occurrence of the same kind for the same
        field is never silenced by an old dismissal. Not valid for
        "race_will_refresh" (it's a quiet note, not a dismissable card) - 404.
      responses:
        "204":
          description: Dismissed.
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Athletes & zones]
      summary: Undo a dismiss
      operationId: undoDismissThresholdSuggestion
      responses:
        "204":
          description: Reverted.
  /v1/athletes/{id}/zones:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes & zones]
      summary: List zone sets
      operationId: listZones
      description: The four zone sets — heart_rate, bike_power, run_power, pace — with computed boundaries.
      responses:
        "200":
          description: Zone sets.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ZoneSet"
  /v1/athletes/{id}/zones/{type}:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum: [heart_rate, bike_power, run_power, pace]
    put:
      tags: [Athletes & zones]
      summary: Replace a zone set
      operationId: replaceZoneSet
      description: >
        Replaces the boundaries for one zone type. Percentages are relative to the
        type's reference threshold.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [zones]
              properties:
                zones:
                  type: array
                  items:
                    $ref: "#/components/schemas/Zone"
      responses:
        "200":
          description: Zone set replaced.
          content:
            application/json:
              schema:
                type: object
                properties:
                  type: { type: string }
                  reference: { type: number }
                  updated: { type: boolean }
  /v1/athletes/{id}/fitness:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Athletes & zones]
      summary: Fitness trend (CTL/ATL/TSB)
      operationId: getFitnessTrend
      description: >
        Returns the daily performance-management series — fitness (CTL), fatigue
        (ATL), and form (TSB) — that drives the dashboard training-load chart.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: A daily fitness series.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FitnessPoint"
...
  /v1/coach/athletes/{id}:
    parameters:
      - $ref: "#/components/parameters/AthleteId"
    get:
      tags: [Coaching]
      summary: Athlete detail
      operationId: getCoachAthlete
      description: A single athlete's training summary, scoped to the coach relationship.
      responses:
        "200":
          description: Athlete training summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachAthlete"
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://api.cadence.cc/oauth/authorize
          tokenUrl: https://api.cadence.cc/oauth/token
          scopes:
            activities:read: Read activities and streams
            activities:write: Upload and edit activities
            workouts:write: Create and edit workouts
            calendar:write: Schedule workouts
            coach: Access an athlete roster
            gear:write: Manage gear
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AthleteId:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Athlete:
      allOf:
        - $ref: "#/components/schemas/User"
      description: Deprecated alias of User, kept for compatibility.
...
    ZoneSet:
      type: object
      properties:
        type: { type: string, enum: [heart_rate, bike_power, run_power, pace] }
        reference: { type: number, description: The threshold the zones are computed from. }
        zones:
          type: array
          items:
            $ref: "#/components/schemas/Zone"
    Zone:
      type: object
      required: [name, low_pct, high_pct]
      properties:
        name: { type: string, example: Z2 Endurance }
        low_pct: { type: number, example: 56 }
        high_pct: { type: number, example: 75 }
...
    ThresholdSummaryEntry:
      type: object
      description: >
        One field's current state - the best qualifying effort within the trailing
        threshold_window_days, so a value can drop as an old best effort ages out,
        not just rise. A manually-entered value (see PATCH /v1/athletes/{id})
        functions the same way from the moment it's set.
      properties:
        value:
          oneOf: [{ type: number }, { type: string }, { type: "null" }]
          description: Watts for ftp/critical_run_power, "mm:ss" for threshold_pace. Null if nothing has ever qualified.
          example: 265
        previous_value:
          oneOf: [{ type: number }, { type: string }, { type: "null" }]
          example: 262
        source_activity_id:
          type: [string, "null"]
          description: Null for a manually-entered value, not sourced from any activity.
          example: act_9f2a7c
        effective_from: { type: [string, "null"], format: date }
        stale:
          type: boolean
          description: >
            The current entry's source has aged out of the trailing window with
            nothing yet to replace it - surfaced for a manual refresh (POST
            /v1/athletes/{id}/thresholds/refresh), never silently corrected. True
            with no entry at all.
          example: false
    ThresholdHistoryEntry:
      type: object
      description: One ledger row - a value that was effective from a given date until superseded.
      properties:
        value:
          oneOf: [{ type: number }, { type: string }]
          example: 265
        source_activity_id: { type: [string, "null"], example: act_9f2a7c }
        effective_from:
          type: string
          format: date
          description: The qualifying activity's own date.
        current_from:
          type: string
          format: date
          description: >-
            The date this row actually became the recorded current value - can differ from
            effective_from when this row only won later, via an earlier better row aging out of
            its trailing window.
    ThresholdSuggestion:
      type: object
      description: >
        One pending Threshold suggestion - shape varies by `kind`. See GET
        /v1/athletes/{id}/threshold-suggestions.
      required: [id, field, kind]
      properties:
        id:
          type: string
          description: "field:kind:key - stable for this occurrence's lifetime."
          example: "ftp:rejected:act_123"
        field: { type: string, enum: [ftp, critical_run_power, threshold_pace] }
        kind: { type: string, enum: [rejected, upcoming_drop, race_will_refresh] }
        current:
          oneOf: [{ type: number }, { type: string }]
          description: The athlete's current value (absent for race_will_refresh). Watts, or "M:SS" for threshold_pace.
          example: 265
        proposed:
          oneOf: [{ type: number }, { type: string }, { type: "null" }]
          description: >
            rejected: the candidate's implied value. upcoming_drop: the successor
            that takes over, or null if the value would simply go stale with
            nothing behind it. Absent for race_will_refresh.
          example: 297
        delta_pct:
          type: [number, "null"]
          description: Unsigned percent change between current and proposed. Absent for race_will_refresh.
          example: 12.1
        activity_id:
          type: [string, "null"]
          description: >
            rejected: the candidate activity. upcoming_drop: the expiring
            entry's own source activity (null if it was a manual entry). Absent
            for race_will_refresh.
          example: act_123
        activity_date:
          type: string
          format: date
          description: rejected/upcoming_drop only - the above activity's own date.
        implied_from:
          type: object
          description: >
            rejected only - the raw effort behind `proposed`, before any
            multiplier (e.g. FTP's 20-min-test x0.95). Equal to `proposed` for a
            field/method with no multiplier step (critical_run_power,
            threshold_pace, and ftp under ftp_calculation_method=sixty_min_direct).
          properties:
            window: { type: string, example: 20min }
            value: { type: number, example: 313 }
        expiry_date:
          type: string
          format: date
          description: upcoming_drop/race_will_refresh only - when the current value ages out.
        days_left:
          type: integer
          description: upcoming_drop/race_will_refresh only - recomputed daily from the cached expiry_date, not itself cached stale.
          example: 12
        race:
          type: object
          description: race_will_refresh only.
          properties:
            id: { type: string }
            name: { type: string }
            date: { type: string, format: date }
...